+++ src/collective/documentgenerator/tests/test_utils.py
@@ -3,7 +3,11 @@ import copy
 import os
 from os import getenv, rmdir
 
+from collective.documentgenerator.utils import update_oo_config
+from plone.api.portal import get_registry_record
+from zope.interface import Interface
 from zope.interface import Invalid
+from zope.lifecycleevent import Attributes
 from zope.lifecycleevent import modified
 
 import collective.documentgenerator as cdg
@@ -69,7 +73,7 @@ class TestUtils(PODTemplateIntegrationTest):
         # don't replace file when not same os file but template is modified
         # We change template file
         test_template.odt_file.data = mgodt
-        modified(test_template)
+        modified(test_template, Attributes(Interface, 'odt_file'))
         self.assertEqual(test_template.initial_md5, mcodt_md5)
         self.assertNotEqual(test_template.style_modification_md5, test_template.current_md5)
         self.assertTrue(test_template.has_been_modified())
@@ -126,3 +130,35 @@ class TestUtils(PODTemplateIntegrationTest):
             os.environ["CUSTOM_TMP"] = initial_custom_tmp
         else:
             del os.environ["CUSTOM_TMP"]
+
+    def test_update_oo_config(self):
+        original_oo_server = getenv("OO_SERVER")
+        original_oo_port = getenv("OO_PORT")
+        original_uno = getenv("PYTHON_UNO")
+
+        lo = "libreofficetest"
+        port = "10422"
+        uno = "/test/fake/python"
+
+        os.environ['OO_SERVER'] = lo
+        os.environ['OO_PORT'] = port
+        os.environ['PYTHON_UNO'] = uno
+
+        update_oo_config()
+
+        oo_server = get_registry_record('collective.documentgenerator.browser.controlpanel.'
+                                        'IDocumentGeneratorControlPanelSchema.oo_server')
+        self.assertEqual(oo_server, lo)
+        oo_port = get_registry_record('collective.documentgenerator.browser.controlpanel.'
+                                      'IDocumentGeneratorControlPanelSchema.oo_port')
+        self.assertEqual(oo_port, int(port))
+        uno_path = get_registry_record('collective.documentgenerator.browser.controlpanel.'
+                                       'IDocumentGeneratorControlPanelSchema.uno_path')
+        self.assertEqual(uno_path, uno)
+
+        if original_oo_server:
+            os.environ['OO_SERVER'] = original_oo_server
+        if original_oo_port:
+            os.environ['OO_PORT'] = original_oo_port
+        if original_uno:
+            os.environ['PYTHON_UNO'] = original_uno