+++ src/collective/documentgenerator/utils.py
@@ -1,19 +1,26 @@
 # -*- coding: utf-8 -*-
+from appy.bin.odfclean import Cleaner
+from collective.documentgenerator import _
+from imio.helpers.security import fplog
+from plone import api
+from plone.namedfile.file import NamedBlobFile
+from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.utils import safe_unicode
+from zope import i18n
+from zope.component import getMultiAdapter
+from zope.component.hooks import getSite
+from zope.component.hooks import setSite
+from zope.interface import Interface
+from zope.interface import Invalid
+from zope.lifecycleevent import Attributes
+from zope.lifecycleevent import modified
+
 import hashlib
 import logging
 import os
 import re
 import tempfile
 
-from plone import api
-from Products.CMFCore.utils import getToolByName
-from Products.CMFPlone.utils import safe_unicode
-from zope import i18n
-from zope.component import getMultiAdapter
-from zope.interface import Invalid
-from zope.lifecycleevent import modified
-
-from collective.documentgenerator import _
 
 logger = logging.getLogger('collective.documentgenerator')
 
@@ -71,7 +78,7 @@ def update_templates(templates, profile='', force=False):
             obj.initial_md5 = new_md5
             obj.style_modification_md5 = new_md5
             obj.odt_file.data = data
-            modified(obj)
+            modified(obj, Attributes(Interface, 'odt_file'))
             ret.append((ppath, ospath, 'replaced'))
     return ret
 
@@ -106,10 +113,10 @@ def ulocalized_time(date, long_format=None, time_only=None, custom_format=None,
         plone = getMultiAdapter((context, request), name=u'plone')
         formatted_date = plone.toLocalizedTime(date, long_format, time_only)
     else:
-        from Products.CMFPlone.i18nl10n import (monthname_msgid,
-                                                monthname_msgid_abbr,
-                                                weekdayname_msgid,
-                                                weekdayname_msgid_abbr)
+        from Products.CMFPlone.i18nl10n import monthname_msgid
+        from Products.CMFPlone.i18nl10n import monthname_msgid_abbr
+        from Products.CMFPlone.i18nl10n import weekdayname_msgid
+        from Products.CMFPlone.i18nl10n import weekdayname_msgid_abbr
         if request is None:
             portal = api.portal.get()
             request = portal.REQUEST
@@ -145,20 +152,29 @@ def remove_tmp_file(filename):
         logger.warn("Could not remove temporary file at {0}".format(filename))
 
 
-def update_oo_config(key='oo_port'):
+def update_oo_config():
     """ Update config following buildout var """
-    var = {'oo_port': 'OO_PORT', 'uno_path': 'PYTHON_UNO'}
-    full_key = 'collective.documentgenerator.browser.controlpanel.IDocumentGeneratorControlPanelSchema.{}'.format(key)
-    configured_oo_option = api.portal.get_registry_record(full_key)
-    new_oo_option = type(configured_oo_option)(os.getenv(var.get(key, 'NO_ONE'), ''))
-    if new_oo_option and new_oo_option != configured_oo_option:
-        api.portal.set_registry_record(full_key, new_oo_option)
+    key_template = 'collective.documentgenerator.browser.controlpanel.IDocumentGeneratorControlPanelSchema.{}'
+    var = {'oo_server': 'OO_SERVER', 'oo_port': 'OO_PORT', 'uno_path': 'PYTHON_UNO'}
+    for key in var.keys():
+        full_key = key_template.format(key)
+        configured_oo_option = api.portal.get_registry_record(full_key)
+        env_value = os.getenv(var.get(key, 'NO_ONE'), None)
+        if env_value:
+            new_oo_option = type(configured_oo_option)(os.getenv(var.get(key, 'NO_ONE'), ''))
+            if new_oo_option and new_oo_option != configured_oo_option:
+                api.portal.set_registry_record(full_key, new_oo_option)
+    logger.info("LibreOffice configuration updated for " + getSite().getId())
+
+
+def update_oo_config_after_bigbang(event):
+    setSite(event.object)
+    update_oo_config()
 
 
 def get_site_root_relative_path(obj):
     return "/" + '/'.join(
         getToolByName(obj, 'portal_url').getRelativeContentPath(obj)
-
     )
 
 
@@ -168,7 +184,7 @@ def temporary_file_name(suffix=''):
         os.mkdir(tmp_dir)
     return tempfile.mktemp(suffix=suffix, dir=tmp_dir)
 
-
+  
 def get_oo_port_list():
     """
     @return the LibreOffice ports numbers to use as a list of int.
@@ -177,3 +193,40 @@ def get_oo_port_list():
         'collective.documentgenerator.browser.controlpanel.IDocumentGeneratorControlPanelSchema.oo_port_list'
     )
     return [port['oo_port'] for port in oo_port_list]
+
+
+def create_temporary_file(initial_file=None, base_name=''):
+    tmp_filename = temporary_file_name(suffix=base_name)
+    # create the file in any case
+    with open(tmp_filename, 'w+') as tmp_file:
+        if initial_file:
+            tmp_file.write(initial_file.data)
+    return tmp_file
+
+
+def clean_notes(pod_template):
+    """ Use appy.pod Cleaner to clean notes (comments). """
+    cleaned = 0
+    odt_file = pod_template.odt_file
+    if odt_file:
+        # write file to /tmp to be able to use appy.pod Cleaner
+        tmp_file = create_temporary_file(odt_file, '-to-clean.odt')
+        cleaner = Cleaner(path=tmp_file.name)
+        cleaned = cleaner.run()
+        if cleaned:
+            manually_modified = pod_template.has_been_modified()
+            with open(tmp_file.name, 'rb') as res_file:
+                # update template
+                result = NamedBlobFile(
+                    data=res_file.read(),
+                    contentType=odt_file.contentType,
+                    filename=pod_template.odt_file.filename)
+            pod_template.odt_file = result
+            if not manually_modified:
+                pod_template.style_modification_md5 = pod_template.current_md5
+            extras = 'pod_template={0} cleaned_parts={1}'.format(
+                repr(pod_template), cleaned)
+            fplog('clean_notes', extras=extras)
+        remove_tmp_file(tmp_file.name)
+
+    return bool(cleaned)