+++ CHANGES.rst
@@ -1,12 +1,81 @@
 Changelog
 =========
 
-3.23 (unreleased)
+3.31 (unreleased)
 -----------------
 
 - Use appy to load balance on pultiple LO server.
   [odelaere]
 
+
+3.30 (2022-05-06)
+-----------------
+
+- Added `iterable_in_columns` (for labels document).
+  [sgeulette]
+- Added `get_relations` in dexterity
+  [sgeulette]
+- Updated readme
+  [Arhell]
+
+3.29 (2022-04-15)
+-----------------
+
+- Method update_oo_config updates all the registry entries for libreoffice server.
+  [odelaere]
+- Added subscriber to update oo config on process start.
+  [odelaere]
+
+3.28 (2022-01-14)
+-----------------
+
+- Added helper `ConfigurablePODTemplate.get_filename` to easily get file filename.
+  [gbastien]
+
+3.27 (2021-12-06)
+-----------------
+
+- Updated metadata version.
+  [sgeulette]
+
+3.26 (2021-11-08)
+-----------------
+
+- Clean notes when creating/editing a `PODTemplate` this way the `search&replace`
+  functionnality works as expected:
+
+  - Added `utils.clean_notes`;
+  - Clean notes is only done if `odt_file` changed and file is updated only if
+    something was cleaned;
+  - Moved function `create_temporary_file` from `events.styles_events` to `utils`.
+
+  [gbastien]
+
+3.25 (2021-09-23)
+-----------------
+
+- Use wrapped context of ProxyObject for __unicode__ method.
+  Revert __repr__ to default implementation.
+  [sdelcourt]
+
+3.24 (2021-09-23)
+-----------------
+
+- Require `beautifulsoup4<4.10` in `setup.py` as starting from 4.10.0,
+  `beautifulsoup4` is no more compatible with `Python2`.
+  [gbastien]
+- Pass parameter `html=True` when calling `Renderer` so content is considered
+  as html when using `xhtml` function in POD template, some pre-processing is
+  applied to ensure that given content is valid xhtml.
+  [gbastien]
+- Use wrapped context of ProxyObject for __repr__ and __str__ methods.
+  [sdelcourt]
+
+3.23 (2021-08-09)
+-----------------
+
+- Allow templates selection on the search result before applying the replace.
+  [sdelcourt]
 
 3.22 (2021-07-16)
 -----------------
@@ -63,7 +132,6 @@ Changelog
 - Use the same default values between controlpanel schema and registry records.
   [odelaere]
 
-
 3.18 (2020-08-18)
 -----------------
 
@@ -105,7 +173,6 @@ Changelog
 
 - Updated the i18n support.
   [macagua]
-
 
 3.14 (2020-04-23)
 -----------------
@@ -261,7 +328,6 @@ Changelog
 - Add hook to provides all helper view to be updated with
   appy renderer
   [sdelcourt]
-
 
 3.0.10 (2018-01-06)
 -------------------
@@ -644,7 +710,6 @@ Changelog
 - Handle case of rendering value of single selection widget.
   [sdelcourt]
 
-
 0.8 (2016-06-03)
 ----------------
 
@@ -652,7 +717,6 @@ Changelog
   to `appy.pod.renderer.Renderer` so private images can be accessed by
   LibreOffice in XHTML fields.
   [gbastien]
-
 
 0.7 (2016-03-22)
 ----------------
@@ -669,7 +733,6 @@ Changelog
   formats are corrects with the kind of file provided.
   [boulch, DieKatze]
 
-
 0.6 (2016-01-21)
 ----------------
 
@@ -680,7 +743,6 @@ Changelog
   The POD template description is now displayed when hovering the POD template title.
   [gbastien]
 
-
 0.5 (2015-12-02)
 ----------------
 
@@ -689,7 +751,6 @@ Changelog
   the `evaluate` method.
   [gbastien]
 
-
 0.4 (2015-12-02)
 ----------------
 
@@ -697,7 +758,6 @@ Changelog
   [gbastien]
 - Take into account the `oo_port` paramater defined in the registry.
   [gbastien]
-
 
 0.3 (2015-09-30)
 ----------------
@@ -711,7 +771,6 @@ Changelog
 - Refactor the generation view to pass the arguments `pod_template` and `output_format`
   directly to the view call or its methods.
   [gbastien, sdelcourt]
-
 
 0.2 (2015-09-22)
 ----------------
@@ -730,13 +789,6 @@ Changelog
   expression is actually evaluated instead of the pod_template and `template`
   is the pod_template
   [gbastien]
-
-
-0.1 (2015-07-17)
-----------------
-
-- Nothing changed yet.
-
 
 0.1 (2015-07-17)
 ----------------